M2-501: Move scroll-related logic into component
