+++ src/shared/ui/index.tsx
@@ -37,6 +37,7 @@ export { default as Slider } from './Slider';
 export { default as Button } from './Button';
 export { default as DateTimePicker } from './DateTimePicker';
 export { default as ScrollButton } from './ScrollButton';
+export { default as ScrollableContent } from './ScrollableContent';
 
 export {
   StatusBar,
